Chapter-15/requirements.txt
openai-agents[voice]
sounddevice
numpy
rich
python-dotenv